requirements.txt
graphiti-core
graphiti-core[google-genai]
python-dotenv
google-generativeai
tavily-python
langchain>=0.0.300
langgraph>=0.1.0
pydantic>=1.10